// src/modules/venta-producto/domain/enums/venta-producto.enum.ts
export enum TipoComprobanteVenta {
  BOLETA = 'boleta',
  FACTURA = 'factura',
}

export enum MetodoPagoVenta {
  EFECTIVO = 'efectivo',
  TARJETA = 'tarjeta',
  YAPE = 'yape',
  PLIN = 'plin',
  TRANSFERENCIA = 'transferencia',
}

export enum EstadoVenta {
  COMPLETADA = 'completada',
  ANULADA = 'anulada',
}